code/requirements.txt
numpy < 2.0.0
PyYAML
torch
torchvision
torchgeometry
torchmetrics
tqdm
